% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_classifier.R
\name{xgboost_classifier}
\alias{xgboost_classifier}
\title{XGBoost Classifier}
\usage{
xgboost_classifier(x, formula = NULL, eta = 0.3, gamma = 0,
  max_depth = 6, min_child_weight = 1, max_delta_step = 0,
  grow_policy = "depthwise", max_bins = 16, subsample = 1,
  colsample_bytree = 1, colsample_bylevel = 1, lambda = 1,
  alpha = 0, tree_method = "auto", sketch_eps = 0.03,
  scale_pos_weight = 1, sample_type = "uniform",
  normalize_type = "tree", rate_drop = 0, skip_drop = 0,
  lambda_bias = 0, tree_limit = 0, num_round = 1, num_workers = 1,
  nthread = 1, use_external_memory = FALSE, silent = 0,
  custom_obj = NULL, custom_eval = NULL, missing = NULL, seed = 0,
  timeout_request_workers = 30 * 60 * 1000, checkpoint_path = "",
  checkpoint_interval = -1, objective = "reg:linear",
  base_score = 0.5, train_test_ratio = 1,
  num_early_stopping_rounds = 0, objective_type = "classification",
  eval_metric = NULL, maximize_evaluation_metrics = FALSE,
  num_class = NULL, base_margin_col = NULL, thresholds = NULL,
  weight_col = NULL, features_col = "features", label_col = "label",
  prediction_col = "prediction", probability_col = "probability",
  raw_prediction_col = "rawPrediction",
  uid = random_string("xgboost_classifier_"), ...)
}
\description{
Foo bar.
}
